requirements.txt
fastapi
uvicorn
opentelemetry-api==1.33.0 # version lock imp
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp-proto-http==1.33.0 #version lock imp
pydantic
pydantic-settings
requests
starlette